Add further tests for orders_controller and correct some abilities checks.
A user could have potentially gotten information about an unrelated order by
attempting to edit the order. Though the user could not update the order, he/she
could view its contents within the edit form.

Further tests are needed. In the next section, we should test not only the
ability of users to update an order, but also the ability to update certain
attributes like the organization and due date.